Casi Listo

+++ client/src/Components/Admin/AddProduct/ProductCardAdmin/ProductCardAdmin.jsx
@@ -6,9 +6,12 @@ import { useNavigate } from 'react-router-dom';
 
 import './ProductCardAdmin.css';
 import { useDispatch } from 'react-redux';
-import { deleteProduct } from '../../../../Redux/actions/actions';
+import {
+  deleteProduct,
+  restoreProduct,
+} from '../../../../Redux/actions/actions';
 
-function ProductCardAdmin({ data }) {
+function ProductCardAdmin({ data, isDeleted }) {
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
@@ -17,7 +20,11 @@ function ProductCardAdmin({ data }) {
   };
 
   const onDelete = () => {
-    dispatch(deleteProduct(data.id, data.type));
+    if (isDeleted) {
+      dispatch(deleteProduct(data.id, data.type));
+    } else {
+      dispatch(restoreProduct(data.id, data.type));
+    }
   };
 
   return (
@@ -34,7 +41,7 @@ function ProductCardAdmin({ data }) {
         <Button onClick={isType} variant="secondary">
           <PencilSquare />
         </Button>
-        <button onClick={onDelete}>Borrar</button>
+        <button onClick={onDelete}>{isDeleted ? 'Borrar' : 'Restaurar'}</button>
       </Card.Body>
     </Card>
   );
